Title: 합과 차의 나눗셈, Time: 165ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240508/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -7,7 +7,7 @@ public class Main {
         int a = sc.nextInt();
         int b = sc.nextInt();
 
-        System.out.printf("%.2f",(a+b)/(a-b));
+        System.out.printf("%.2f",(double)(a+b)/(a-b));
 
 
     }
